clean up queries
